+++ packages/browser/src/extensions/tracing-headers.ts
@@ -38,10 +38,16 @@ export class TracingHeaders {
 
     private _startCapturing = () => {
         if (isUndefined(this._restoreXHRPatch)) {
-            assignableWindow.__PosthogExtensions__?.tracingHeadersPatchFns?._patchXHR(this._instance.sessionManager)
+            assignableWindow.__PosthogExtensions__?.tracingHeadersPatchFns?._patchXHR(
+                this._instance.get_distinct_id(),
+                this._instance.sessionManager
+            )
         }
         if (isUndefined(this._restoreFetchPatch)) {
-            assignableWindow.__PosthogExtensions__?.tracingHeadersPatchFns?._patchFetch(this._instance.sessionManager)
+            assignableWindow.__PosthogExtensions__?.tracingHeadersPatchFns?._patchFetch(
+                this._instance.get_distinct_id(),
+                this._instance.sessionManager
+            )
         }
     }
 }